+++ packages/terra-scroll/src/terra-dev-site/doc/scroll/Scroll.doc.jsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../docs/README.md';
-import { name } from '../../../package.json';
+import ReadMe from '../../../../docs/README.md';
+import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
 import ScrollSrc from '!raw-loader!terra-scroll/src/Scroll';
 
 // Example Files
-import ScrollVertical from './example/ScrollVertical';
-import ScrollVerticalSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/ScrollVertical';
-import ScrollHorizontal from './example/ScrollHorizontal';
-import ScrollHorizontalSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/ScrollHorizontal';
+import ScrollVertical from '../example/ScrollVertical';
+import ScrollVerticalSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ScrollVertical';
+import ScrollHorizontal from '../example/ScrollHorizontal';
+import ScrollHorizontalSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/ScrollHorizontal';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
