+++ packages/terra-search-field/src/terra-dev-site/doc/search-field/SearchField.doc.jsx
@@ -1,31 +1,31 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../docs/README.md';
-import { name } from '../../../package.json';
+import ReadMe from '../../../../docs/README.md';
+import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
-import SearchFieldSrc from '!raw-loader!../../../src/SearchField';
+import SearchFieldSrc from '!raw-loader!../../../../src/SearchField';
 
 // Example Files
-import SearchFieldBasic from './example/SearchFieldBasic';
-import SearchFieldBasicSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldBasic';
-import SearchFieldPlaceholder from './example/SearchFieldPlaceholder';
-import SearchFieldPlaceholderSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldPlaceholder';
-import SearchFieldMinimumLength from './example/SearchFieldMinimumLength';
-import SearchFieldMinimumLengthSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldMinimumLength';
-import SearchFieldDelayed from './example/SearchFieldDelayed';
-import SearchFieldDelayedSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldDelayed';
-import SearchFieldDefaultValue from './example/SearchFieldDefaultValue';
-import SearchFieldDefaultValueSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldDefaultValue';
-import SearchFieldDisabled from './example/SearchFieldDisabled';
-import SearchFieldDisabledSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldDisabled';
-import SearchFieldBlock from './example/SearchFieldBlock';
-import SearchFieldBlockSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldBlock';
-import SearchFieldDisableAutoSearch from './example/SearchFieldDisableAutoSearch';
-import SearchFieldDisableAutoSearchSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldDisableAutoSearch';
-import SearchFieldFilterNumeric from './example/SearchFieldFilterNumeric';
-import SearchFieldFilterNumericSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SearchFieldFilterNumeric';
+import SearchFieldBasic from '../example/SearchFieldBasic';
+import SearchFieldBasicSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldBasic';
+import SearchFieldPlaceholder from '../example/SearchFieldPlaceholder';
+import SearchFieldPlaceholderSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldPlaceholder';
+import SearchFieldMinimumLength from '../example/SearchFieldMinimumLength';
+import SearchFieldMinimumLengthSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldMinimumLength';
+import SearchFieldDelayed from '../example/SearchFieldDelayed';
+import SearchFieldDelayedSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldDelayed';
+import SearchFieldDefaultValue from '../example/SearchFieldDefaultValue';
+import SearchFieldDefaultValueSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldDefaultValue';
+import SearchFieldDisabled from '../example/SearchFieldDisabled';
+import SearchFieldDisabledSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldDisabled';
+import SearchFieldBlock from '../example/SearchFieldBlock';
+import SearchFieldBlockSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldBlock';
+import SearchFieldDisableAutoSearch from '../example/SearchFieldDisableAutoSearch';
+import SearchFieldDisableAutoSearchSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldDisableAutoSearch';
+import SearchFieldFilterNumeric from '../example/SearchFieldFilterNumeric';
+import SearchFieldFilterNumericSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SearchFieldFilterNumeric';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
