+++ packages/terra-section-header/src/terra-dev-site/doc/section-header/ChangeLog.doc.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DocTemplate from 'terra-doc-template';
+import ChangeLog from '../../../../CHANGELOG.md';
+
+const DocPage = () => (
+  <DocTemplate
+    packageName="terra-section-header"
+    srcPath="https://github.com/cerner/terra-core/tree/master/packages/terra-section-header"
+    readme={ChangeLog}
+  />
+);
+
+export default DocPage;
